u-boot-zynq-uenv.bb: Handle IMAGE_BOOT_FILES wildcard patterns
Handle IMAGE_BOOT_FILES having wildcard patterns for dtbs. This requires
that the recipe depend on the deployment of the dtbs in order for the
correct file names to be expanded.

This change also improves the parsing of the IMAGE_BOOT_FILES variable
and splits out the python functions for doing this into a separate
image-wic-utils.bbclass file.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>