microzed-zynq7: Change to using device tree from kernel
Change the microzed board to use the device tree from the kernel. This
device tree was added in in kernel v4.10 and was pulled into the
meta-xilinx layer to resolve some usb issues in commit 8e01359125.

With all kernel recipes being at or beyond v4.10 using the kernel
source is straight forward and removes any issues caused by differences
between the in kernel and the copied version.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>