machine-xilinx-default.inc: Add dtb files for IMAGE_BOOT_FILES
Add all available dtb sources to the IMAGE_BOOT_FILES variable. For
device-tree recipe generated dtbs the files available are only known
after they files are deployed, so a wildcard pattern is used for these
files. Note that this pattern appears before the kernel device trees due
to the preference to use custom non-kernel device trees where available.
This ordering is needed so that recipes like u-boot-zynq-uenv can pick
the first device tree to select as the default.

The kernel device trees are specifically selected based on the value of
KERNEL_DEVICETREE, this avoid the duplication of kernel image type
prefixed files along side having the actual files that the kernel
builds.

Additionally remove all instances of "${MACHINE}.dtb" as they no longer
need to be specified and are incorrect due to the nesting of the files
in the `devicetree` directory.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>